pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ytcli"
version = "0.3.0"
description = "Multi-account YouTube playlist/subscription manager"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client",
    "google-auth-oauthlib",
    "PyYAML"
]

[tool.setuptools]
packages = ["ytcli"]

